+++ mcp_tools/web_app_security/enhanced_web_app_security.py
@@ -5,71 +5,6 @@ from typing import Dict, Any, Optional
 def register_enhanced_web_app_security_tools(mcp, hexstrike_client, logger, HexStrikeColors):
 
     @mcp.tool()
-    def uro_url_filtering(urls: str, whitelist: str = "", blacklist: str = "",
-                         additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute uro for filtering out similar URLs.
-
-        Args:
-            urls: URLs to filter
-            whitelist: Whitelist patterns
-            blacklist: Blacklist patterns
-            additional_args: Additional uro arguments
-
-        Returns:
-            Filtered URL results with duplicates removed
-        """
-        data = {
-            "urls": urls,
-            "whitelist": whitelist,
-            "blacklist": blacklist,
-            "additional_args": additional_args
-        }
-        logger.info("🔍 Starting uro URL filtering")
-        result = hexstrike_client.safe_post("api/tools/uro", data)
-        if result.get("success"):
-            logger.info("✅ uro URL filtering completed")
-        else:
-            logger.error("❌ uro URL filtering failed")
-        return result
-
-    @mcp.tool()
-    def burpsuite_scan(project_file: str = "", config_file: str = "", target: str = "", headless: bool = False, scan_type: str = "", scan_config: str = "", output_file: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Burp Suite with enhanced logging.
-
-        Args:
-            project_file: Burp project file path
-            config_file: Burp configuration file path
-            target: Target URL
-            headless: Run in headless mode
-            scan_type: Type of scan to perform
-            scan_config: Scan configuration
-            output_file: Output file path
-            additional_args: Additional Burp Suite arguments
-
-        Returns:
-            Burp Suite scan results
-        """
-        data = {
-            "project_file": project_file,
-            "config_file": config_file,
-            "target": target,
-            "headless": headless,
-            "scan_type": scan_type,
-            "scan_config": scan_config,
-            "output_file": output_file,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting Burp Suite scan")
-        result = hexstrike_client.safe_post("api/tools/burpsuite", data)
-        if result.get("success"):
-            logger.info(f"✅ Burp Suite scan completed")
-        else:
-            logger.error(f"❌ Burp Suite scan failed")
-        return result
-
-    @mcp.tool()
     def zap_scan(target: str = "", scan_type: str = "baseline", api_key: str = "", daemon: bool = False, port: str = "8090", host: str = "0.0.0.0", format_type: str = "xml", output_file: str = "", additional_args: str = "") -> Dict[str, Any]:
         """
         Execute OWASP ZAP with enhanced logging.
