+++ mcp_tools/web_scan/burpsuite.py
@@ -1,0 +1,41 @@
+# mcp_tools/web_scan/burpsuite.py
+
+from typing import Dict, Any
+
+def register_burpsuite_tool(mcp, hexstrike_client, logger):
+
+    @mcp.tool()
+    def burpsuite_scan(project_file: str = "", config_file: str = "", target: str = "", headless: bool = False, scan_type: str = "", scan_config: str = "", output_file: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Burp Suite with enhanced logging.
+
+        Args:
+            project_file: Burp project file path
+            config_file: Burp configuration file path
+            target: Target URL
+            headless: Run in headless mode
+            scan_type: Type of scan to perform
+            scan_config: Scan configuration
+            output_file: Output file path
+            additional_args: Additional Burp Suite arguments
+
+        Returns:
+            Burp Suite scan results
+        """
+        data = {
+            "project_file": project_file,
+            "config_file": config_file,
+            "target": target,
+            "headless": headless,
+            "scan_type": scan_type,
+            "scan_config": scan_config,
+            "output_file": output_file,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting Burp Suite scan")
+        result = hexstrike_client.safe_post("api/tools/burpsuite", data)
+        if result.get("success"):
+            logger.info(f"✅ Burp Suite scan completed")
+        else:
+            logger.error(f"❌ Burp Suite scan failed")
+        return result
